fix: fetch fails in Safari 14 as requests from background script are considered as CORS
* Fetch is now done in content script
* Moved TokenSigning script to separate files to clean up content.ts
