+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "web-eid-webextension",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "",
   "main": "src/index.js",
   "scripts": {
